Added publican-redhat-db5 brand.
The publican-redhat-db5 brand is intended for use with Red Hat branded
books that make use of Publican's experimental support for DocBook 5.

Initially the brand largely relies on fallbacks to Publican's XSL files
which are by and large written for transforming DocBook 4. In the long
run it is intended that these will be rewritten/updated.

This initial commit also includes DocBook 5 compatible versions of the
common content and some minor CSS tweaks. Note that the conversion of
the common content to DocBook 5 is likely to have resulted in *some* of
the translation files being out of alignment at this time. Care is
recommended when using this brand with a book that either has or will
have translations.